pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-swarm"
version = "0.1.0"
description = "Agentic Task OS with HITL workflow and Meridian_Aux plugin"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "data_swarm maintainers"}]
dependencies = [
  "pyyaml>=6.0",
  "openai>=1.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]

[project.scripts]
data-swarm = "data_swarm.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["D203", "D212"]

[tool.ruff.lint.pydocstyle]
convention = "google"
